Improve logic that decides when to swap delegates while a widget is being dragged
So that you don't have to drag too much a widget to get a swap to take place.

Eg: have two widgets, a 4-rows-tall one on the top and a one-row-tall at the
bottom. Then drag the big top one until it swap places with the bottom one.
Without this patch you would have to drag quite a lot.