WidgetGrid: Proper animations when widgets are added or removed etc
To make that possible a major refactoring was needed. Had to ditch
Column{} and go for a full custom implementation. Tried to use ListView
but it didn't work well as widget additions and removals also affected
sizes and positions of other items in the list and ListView wasn't designed
for that. There's also the case that when there's a single item in the list
it should be centered and have a certain size, and I would have to torture
ListView quite a lot to get this implemented.

Now WidgetListModel is an integral part of WidgetGrid and thus was moved
inside it. The model and the qml code work hand in hand.

Removed all header dependencies from WidgetListModel and put it into a plugin
so that WidgetGrid could still be used in qml tests and be fed with a model
fully written in qml.

Also removed appman dependencies in ApplicationInfo.