Make it impossible for apps to mess up widget decorations and borders
Application windows are not clipped exactly on the rounded widget
borders. And now the window is put next to the orange ribbon
instead of in front of it, which makes it impossible to cover
the widget orange ribbon as well.

Additionally, sysui no longer changes the width of application
windows at all. Just like it was already being done for the height,
exposedRect is used to inform the application of the area of its
fullscreen window that is actually being displayed.