Update the settings.qface file
Merge old settings under one interface UISettings, and add the InstrumentCluster interface.
App halfway up-to-date. No need for separate service implementations in the server end
as the server does not need to save the values persistently for now. There are some problems
with using enums in the interface, so ill have to see later if i could add those later (needed
for the drivetrain and theme)