initial about page with marketing and monitor view
This requires still some review by design and the monitor view are
not really beautiful. We could use the QtCharts for rendering more
advanced graphs. Also there is no network monitor available.

You can launch the about page by clicking the Qt icon. it will be closed
on any click on the overlay.