Remove the default width and height from the Tool Component
By default Tool is using the dimensions of the passed symbol and the
Label, so the default is overwritten.

In case of an Translation the bindings are reevaluated and because of that
the widht and height are set, which leads to broken behavior. As we don't
want to hardcode every Tool, just removing the default fixes the behavior.