[SysUI] Corrected animation in ApplicationWidget
In windowSlot Item leftMargin.

The difference is mostly noticed in MusicApp. With smoothed animation
the content in full screen (tool bar & music list) it is moving
a bit from right to left when opening, in the meantime it looks
more steady with just number animation. Agreed with designer that
NumberAnimation is what we need here.

Task-number: AUTOSUITE-214
Change-Id: Ic03c55cbb950828870971cb9c7fb8ec077952849
Reviewed-by: Daniel d'Andrada <daniel.dandrada@luxoft.com>