[cluster] fix music and maps freezing w/o cluster window
Add a new settings property InstrumentCluster.available
and use it in apps to decide when to create and display
their secondary windows.

Also reacts to runtime changes, e.g. when the secondary cluster
window gets closed by the user.

Can be tested by setting "showCluster" in the am-config.yaml
to "no".

Task-number: AUTOSUITE-486
Change-Id: Ia6bdc65aa6dc1d4211e89879e48aafaca7c54aa6
Reviewed-by: Bramastyo Harimukti Santoso <bramastyo.harimukti.santoso@pelagicore.com>