Use the iconTheme config options from applicationmanager
Previously a ProxyTheme class was used in the neptunestyleplugin, where
a path was returned which was hardcoded to the INSTALL_PREFIX. This
doesn't work when neptune is run from the build-folder and wasn't
installed previously or when installed from the online installer
as the theme directory needed to be known at build-time.

The new solution is to use the iconThemeName and iconThemeSearchPaths
configuration options of application-manager and get rid of the
ProxyTheme completely. This also makes sure neptune3 will compile
with QQC2 from the dev branch, which doesn't offer the ProxyTheme
anymore.

Task-number: QTAUTO-1072
Change-Id: I2d9c4f888c6430613d4a56480b37a364517c44e9
Reviewed-by: Lukáš Tinkl <ltinkl@luxoft.com>
Reviewed-by: Bramastyo Harimukti Santoso <bramastyo.harimukti.santoso@pelagicore.com>