[config] Add a special server.conf for windows
This is needed as on windows clients cannot connect to a server
with the url 0.0.0.0.
On windows we now use a different server.conf using 127.0.0.1
as the url for server and client. This has the limitation that
the android remote app can not connect to a server running on this
localhost only address.

Task-number: AUTOSUITE-546
Change-Id: Ib07406373d533f185dde3f5a38376bf68f289055
Reviewed-by: Bramastyo Harimukti Santoso <bramastyo.harimukti.santoso@pelagicore.com>