[sysui] registerCompositorView() is no longer in appman's public API
It's now called internally automatically.

The drawback is that it's only called once the toplevel window is made
visible for the first time, which is what triggers creation of its
platform window and backing surface.

So the logic that to make InstrumentClusterWindow visible only when a
instrument cluster app window is already available won't work anymore,
As the InstrumentClusterWindow has to be already propertly registered in
order to receive any client window for composition.

On the other hand there was no use case for that visibility behavior anyway.

Change-Id: Iaca5ac677a23c8d118d0752469dd448db8885bdc
Reviewed-by: Nikolay Zamotaev <nzamotaev@luxoft.com>
Reviewed-by: Bramastyo Harimukti Santoso <bramastyo.harimukti.santoso@pelagicore.com>