[sysui] Don't set InstrumentCluster replica value while uninitialized
If the InstrumentCluster replica is still uninitialized when its available
property is set to true, this property value will be silently ignored and
lost once initialization is complete.

You have to wait until the replica is initialized for the property values to
actually work (ie, get sent to server and thus reflected on the replicas in all
client applications).

Change-Id: I239bcd31e98f730a6ee9bbf75403a28932ed3399
Reviewed-by: Bramastyo Harimukti Santoso <bramastyo.harimukti.santoso@pelagicore.com>