Fix neptune3 on android
With the recent import path refactorings the deployment for android got
broken. This fixes the paths for the C++ plugins to let them be deployed
by androiddeployqt all QML only modules are deployed as normal assets
and are imported using appman.

To correctly deploy the controls plugin which has mixed content (QML + C++)
the qmlplugin.prf got improved to copy the needed files to the build
folder, which should also help to start neptune3 from the build-folder.

Change-Id: I6d875702d67c80a0a6f4ce8bde00a62eb1d8abcb
Reviewed-by: Bramastyo Harimukti Santoso <bramastyo.harimukti.santoso@pelagicore.com>
Reviewed-by: Daniel d'Andrada <daniel.dandrada@luxoft.com>