Fix the Companion app on android
We make sure to always build both type of qtivi backends. The production
backend using qtro and the simulation backend. When building neptune3
for android the production backend is not build and running it will
automatically fall back to the simulation there.

With the companion app we build the production backend to be able to
connect to a running neptune3 on a target.

Change-Id: I0aebd34ba730956566db05da3c571a425e9f61b3
Reviewed-by: Bramastyo Harimukti Santoso <bramastyo.harimukti.santoso@pelagicore.com>