[doc] Correctly outline the docs in the generated qhp files
When opening the qhp in assistant or in QtCreator (Contents view) the
Neptune3 content had a lot of broken links and mixed general documentation
with QML type documentation.

This change makes sure to use the same TOC for the "contents" view than
what is used in the index.html and groups the QML type documentation
into a "QML Types" group similar to all other modules.

Change-Id: I129d2e0146312bc929c3bc918ecac7d40ec96ba2
Reviewed-by: Bramastyo Harimukti Santoso <bramastyo.harimukti.santoso@pelagicore.com>