[squish] connect to squish-appman-hook, little clean-up
Appman hook will be soon present, and in order to use it, the AUT must
get the '-c' parameter pointing to the squish-appman-hook.yaml. It
will be located in the install (root) folder, so that is where it
points to. Clean-up was needed to remove unnecessary parts and overcomplicated
sections.

Change-Id: I7ca33b9fa1d57a30fc3d14fb5470e508be641e3e
Reviewed-by: Bramastyo Harimukti Santoso <bramastyo.harimukti.santoso@pelagicore.com>