[squish] add simple calendar app test
A simple calendar app test is added for multi-process
testing with apps. It is an app test that is simple, in
order to have more robust functioning app tests, because movement
(tap and move) is still a problem. It switches between
year, events, and next view.
Changes to code is needed to receive according objectName
values, to identify the buttons.

Change-Id: I11ac62e7cefce7ced2fc96eb442bf205aff9ca0f
Reviewed-by: Bramastyo Harimukti Santoso <bramastyo.harimukti.santoso@pelagicore.com>