[squish] add view check to calendar view button test
The calendar test was too simple: only buttons can be clicked,
but no real check that the view actually changed. This
is now added, and put into a scenario outline.
ObjectNames had to be added to check names, changed index
check alone wouldn't be so meaningful.
To let tests run alone, switching to main app was important
to add.

Change-Id: Icb3636cef1bc98ca592df5cb6f0ceb79456ade71
Reviewed-by: Bramastyo Harimukti Santoso <bramastyo.harimukti.santoso@pelagicore.com>