[squish] fix single/multi process by container switch
For single/multi process the containers for apps are
different, since the apps act on the app itself or
in single mode the center console as starting container
to search in. Context switch for single
process is not needed, and therefore will have no
effect. A internal variable is introduced, using the
number of contexts found to find out if single or multi
process is running.
For an intention and to start tests
accordingly, an environment variable before tests start
is used, called SINGLE_PROCESS, mainly for Linux
systems which by default start in multi process.
Unnecessary and misleading object map files are deleted.

Task-number: AUTOSUITE-796
Change-Id: Ia1e42adee5aac1b9e196c823ad2ccd299b0fef69
Reviewed-by: Bramastyo Harimukti Santoso <bramastyo.harimukti.santoso@pelagicore.com>