[squish] fix dbus/RemoteSettingsServe issue by documentation
Earlier dbus and current settings problem result in
disruptive sub processes that Squish cannot handle.
To solve this, Squish has an ignore option which
excludes selective processes.
The option is set in an additional Squish settings
file, which cannot be included here, but solving the
problem, is to update and read the documentation
which shows how to do that.

Change-Id: I4703258301766e14177d5f6d9a6ed2885acc0528
Reviewed-by: Kavindra Palaraja <kpalaraja@luxoft.com>
Reviewed-by: Bramastyo Harimukti Santoso <bramastyo.harimukti.santoso@pelagicore.com>