[squish] add simple phone widget test
The phone widget has a short list of numbers to
call, mainly it's the same as phone app. Hence
a new test is introduced, common steps of phone
moved to a new file, and a (later to be moved)
general list function had to be added.
From the favorites short view, persons can be called
by giving the name.

Change-Id: I6034fd57d6c557aab49bc7e6f7c53a28baa05942
Reviewed-by: Bramastyo Harimukti Santoso <bramastyo.harimukti.santoso@pelagicore.com>