[squish] fix apps common start to always open via launcher grid
In a common function, apps could start from the
first x items, that are in the launch bar and directly
visible, for other's the grid button would be pressed
and then the app tapped.
To open via full app grid will be the default behavior
since it proves to be more robust, when app seem to
be visible on the launcher but strangely are not
tap/click-able.

Change-Id: I191f954fc4c1526e9b9fc970b50f9e7bfe983171
Reviewed-by: Bramastyo Harimukti Santoso <bramastyo.harimukti.santoso@pelagicore.com>