Remote settings lib has no dependency to QtQuick
Do not link against QtQuick, as it caused linking error with the
remote settings server when using QEMU toolchain on CI.

The linker in QEMU toolchain does not default to --as-needed, causing
lib to have dependency to QtQuick. When the server was linked without
QtQuick, it caused linker to find wrong version of QtQuick from the
sysroot.

Change-Id: I523c62929c06b8977fe0c4fe4e20fd87ea37243c
Reviewed-by: Dominik Holland <dominik.holland@pelagicore.com>
Reviewed-by: Bramastyo Harimukti Santoso <bramastyo.harimukti.santoso@pelagicore.com>