[build-system] Fix in-tree builds
Make the qmlplugin.prf more flexible by always using EXTRA_FILES
instead of having a second special rule for the qmldir file.
In case a qmldir exists it will be automatically added to EXTRA_FILES.
The EXTRA_FILES can now also be part of another folder, which is
needed for the future when qmlplugins are autogenerated as well
and the qmldir files are part of the build folder.

Change-Id: I0b310a19f611e8e5d7cb65de5de9fe06cfaf01e1
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>