[sysui] add warning about switching off autostart for all apps to docs
It is possible to switch off autostart for all apps including system
ones (Climate, Cluster, HUD). On restart this will lead to empty windows,
which might be confusing for the user. So we add warnings to the docs and
remove differentiation between system/non-system apps in the list.

Task-number: AUTOSUITE-1026
Change-Id: I2b37451d68661fa5bad54e91dbce581cd6fe5a8e
Reviewed-by: Kavindra Palaraja <kpalaraja@luxoft.com>