[neptune-cluster-app] cluster as a regular app
This commit introduces the cluster app as client for middle-ware
services, the cluster contains mocked views for limited list of apps.
These apps are supposed to be vies for Map, Music, Vehicle.
This patch doesn't affect neptune3 as app, it only adds new application.

It is built as part of neptune3-ui prj;
independent *.pro will be added later

This commit is the initial, *.yaml configuration handling and languages
will be added in following commits.

Later corresponding parts inside neptune3 will be disabled and
moved inside this subprj.


Task-number: AUTOSUITE-1267
Fixes: AUTOSUITE-1288
Change-Id: I3d4a4519489c2893a0a6849805e48bfedda8dee8
Reviewed-by: Egor Nemtsev <enemtsev@luxoft.com>