[sysui] add processing of icon.source for ToolsColumn component
- ToolsColumn uses ToolButton inside. ToolButton provides two ways of
setting icon image. First one is using icon.name and theme, it is now
implemented in Neptune 3 UI. Another uses icon URL in icon.source.
This commit adds processing of sourceOn and sourceOff roles in the
component's model to define the ToolButton's icon.source property for
the selected and deselected states.

Task-number: AUTOSUITE-1215
Change-Id: I33f3a0e133db81e03464a430d5e2cd0302026f80
Reviewed-by: Kavindra Palaraja <kpalaraja@luxoft.com>