[build-system] Make Neptune3 behave like a Qt module
Start using the qt_parts qmake feature, like other Qt modules are doing
it. This enables the build system to take Qt's configuration into
account and only build examples and tests when Qt was configured to do
so.

Similar to QtIvi and QtApplicationManager the enable-examples and
enable-tests config options can be used to manually enable building these
parts.

Task-number: AUTOSUITE-1308
Change-Id: I7b316a4b5c509a08b97ecc1cc4c645096ffeb8ff
Reviewed-by: Grigorii Zimin <gzimin@luxoft.com>
Reviewed-by: Samuli Piippo <samuli.piippo@qt.io>