[android] Fix deployment of the qtivi plugins
Because of the new AAB of Qt for android, the deployment mechanism has
been changed and the plugin location is now also encoded into the file
name, similar to how it was done with QML plugins already before.

Using qt5LibraryTarget instead of qtLibraryTarget makes sure the prefix
is correct and the plugins are detected correctly by QtIvi.

Change-Id: I880cee19f59c6bdeb7c3c9d3bff01b9762707df0
Task-number: AUTOSUITE-1408
Reviewed-by: Egor Nemtsev <enemtsev@luxoft.com>