[launcher] fix elements arrangement
--add drop area borders (strict area where user can leave elements)
--remove this all "cargo-stuff" from editable grid view, each appbutton
is owned by its own DragHandler and there is no need to change it on any
user's move (we already control enable property).

Behavior that caused this bug:
DragHandler target was by default set to parent, and sometimes there was
a gap between setting target to appbutton, which caused a try to acquire
appbutton's field from qgraphitem object.

Task-number: AUTOSUITE-1435
Change-Id: I09e73929601e522fe297236d8f670d55f48a6ebd
Reviewed-by: Egor Nemtsev <enemtsev@luxoft.com>